+++ app/Repositories/Backend/Place/PlaceRepository.php
@@ -12,6 +12,13 @@ use Illuminate\Database\Eloquent\Model;
 use App\Repositories\Backend\Access\Role\RoleRepository;
 use App\Models\Place\PlaceMedias;
 
+use App\Models\ActivityMedia\Media;
+use App\Models\ActivityMedia\MediaTranslations;
+use App\Models\Access\language\Languages;
+
+use App\Helpers\UrlGenerator;
+
+
 /**
  * Class PlaceRepository.
  */
@@ -115,6 +122,44 @@ class PlaceRepository extends BaseRepository
 
             if ($model->save()) {
 
+                if(!empty($extra['files'])){
+                    
+                    $url = UrlGenerator::GetUploadsUrl();
+                    $i = 0;
+                    foreach ($extra['files'] as $key => $file) {
+                        $extension = $file->extension();
+                        
+                        if(self::validateUpload($extension)){
+                            $new_file_name = time() . $i++ . '_place.' . $file->extension();
+                            $new_path = '/uploads/medias/places/' . $model->id . '/';
+                            $file->storeAs( $new_path , $new_file_name);
+                            
+                            $media = new Media;
+                            $media->url = $url . 'medias/places/' . $model->id . '/' . $new_file_name;
+                            $media->type = Media::TYPE_IMAGE;
+                            $media->save();
+                            
+                            $languages = Languages::all();
+
+                            if(!empty($languages)){
+                                foreach ($languages as $key => $value) {
+                                    $media_trans = new MediaTranslations;
+                                    $media_trans->medias_id = $media->id;
+                                    $media_trans->languages_id = $value->id;
+                                    $media_trans->title = $new_file_name;
+                                    $media_trans->description = "Image";
+                                    $media_trans->save();
+                                }
+                            }
+
+                            $place_media = new PlaceMedias;
+                            $place_media->places_id = $model->id;
+                            $place_media->medias_id = $media->id;
+                            $place_media->save();
+                        }
+                    }
+                }
+
                 if(!empty($extra['medias'])){
                     foreach ($extra['medias'] as $key => $value) {
                        $PlaceMedias = new PlaceMedias;
@@ -185,7 +230,38 @@ class PlaceRepository extends BaseRepository
         $prev_medias = PlaceMedias::where(['places_id' => $id])->get();
         if(!empty($prev_medias)){
             foreach ($prev_medias as $key => $value) {
-                $value->delete();
+                if($value->media->type == null){
+                    $value->delete();
+                }
+            }
+        }
+
+        if(!empty($extra['delete-images'])){
+            $images_arr = explode(',' , $extra['delete-images']);
+            
+            if(!empty($images_arr)){
+                foreach ($images_arr as $key => $value) {
+                    $temp = Media::where(['id' => $value])->first();
+                    
+                    if(!empty($temp)){
+                        if($temp->type == Media::TYPE_IMAGE){
+                            $places_media = PlaceMedias::where(['medias_id' => $temp->id])->first();
+                            $path = storage_path() . DIRECTORY_SEPARATOR . 'uploads' . DIRECTORY_SEPARATOR . 'medias' . DIRECTORY_SEPARATOR . 'places' . DIRECTORY_SEPARATOR;
+                            if(!empty($places_media)){
+                                
+                                $filename = explode('/',$temp->url);
+                                $filename = end($filename);
+                                $path .= $places_media->places_id . DIRECTORY_SEPARATOR;
+                                $path .= $filename;
+                                
+                                if(is_file($path)){
+                                    unlink($path);
+                                }
+                            }
+                            $temp->delete();
+                        }
+                    }
+                }
             }
         }
 
@@ -193,6 +269,44 @@ class PlaceRepository extends BaseRepository
             $check = 1;
 
             if ($model->save()) {
+
+                if(!empty($extra['files'])){
+                    
+                    $url = UrlGenerator::GetUploadsUrl();
+                    $i = 0;
+                    foreach ($extra['files'] as $key => $file) {
+                        $extension = $file->extension();
+                        
+                        if(self::validateUpload($extension)){
+                            $new_file_name = time() . $i++ . '_place.' . $file->extension();
+                            $new_path = '/uploads/medias/places/' . $model->id . '/';
+                            $file->storeAs( $new_path , $new_file_name);
+                            
+                            $media = new Media;
+                            $media->url = $url . 'medias/places/' . $model->id . '/' . $new_file_name;
+                            $media->type = Media::TYPE_IMAGE;
+                            $media->save();
+                            
+                            $languages = Languages::all();
+
+                            if(!empty($languages)){
+                                foreach ($languages as $key => $value) {
+                                    $media_trans = new MediaTranslations;
+                                    $media_trans->medias_id = $media->id;
+                                    $media_trans->languages_id = $value->id;
+                                    $media_trans->title = $new_file_name;
+                                    $media_trans->description = "Image";
+                                    $media_trans->save();
+                                }
+                            }
+
+                            $place_media = new PlaceMedias;
+                            $place_media->places_id = $model->id;
+                            $place_media->medias_id = $media->id;
+                            $place_media->save();
+                        }
+                    }
+                }
 
                 if(!empty($extra['medias'])){
                     foreach ($extra['medias'] as $key => $value) {
@@ -235,5 +349,26 @@ class PlaceRepository extends BaseRepository
 
             throw new GeneralException('Unexpected Error Occured!');
         });
+    }
+
+    public static function validateUpload($extension) {
+        
+        $extension = strtolower($extension);
+
+        switch($extension){
+            case 'jpeg':
+                return true;
+            case 'jpg':
+                return true;
+                break;
+            case 'png':
+                return true;
+                break;
+            case 'gif':
+                return true;
+                break;
+            default:
+                return false;
+        }
     }
 }
